+++ tools/lgtdoc/xml/lgt2md.js
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////////////////////////////////
 //
 //   XML documenting files to Mardown text files conversion script 
-//   Last updated on July 5, 2016
+//   Last updated on October 24, 2018
 //
 //   This file is part of Logtalk <https://logtalk.org/>  
 //   Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -146,7 +146,7 @@ for (files.moveFirst(); !files.atEnd(); files.moveNext()) {
 	if (FSObject.GetExtensionName(file) == "xml") {
 		WScript.Echo("  converting " + file);
 		var md_file = directory + "\\" + FSObject.GetBaseName(file) + ".md";
-		if (file == "directory_index.xml" || file == "entity_index.xml" || file == "predicate_index.xml") {
+		if (file == "library_index.xml" || file == "directory_index.xml" || file == "entity_index.xml" || file == "predicate_index.xml") {
 			xslt = index_xslt;
 		} else {
 			xslt = entity_xslt;